Update DynamicSettingsRootView.swift
Remove warning re: Sigmoid formula. Both logarithmic and sigmoid are imperfect based on recent analysis by Chris Wilson, David Burren, and Theresa Hastings. This statement as is omits logarithmic shortfalls and including logarithmic in the app description only causes unclear confusion for users. At this time it is best to remove this statement until a new formula for both is determined. It may be better to add information in the documentation rather than in the app.