Adjust tap gestures for Adjustments UI
Update Add/Edit OR/TT so tapping labels doesn't do anything, and make it so only tapping the values brings up the picker for that variable. This matches the UI currently in the settings modules.

Also update the "Stop OR/TT" button so tapping anywhere on the button triggers it and not just tapping on the button's text label.