CGM backfill support
This commit adds logic to support backfilled CGM values from CGM
managers. There are a few key corner cases that it handles:
- _Not_ backfilling previously deleted CGM readings
- Deduplication for CGM readings
- Avoiding adding CGM readings when values already exist

The way we handle it is by comparing timestamps from both the
DeletedGlucoseStored and GlucoseStored CoreData tables.

Our logic isn't perfect, because we're only looking at timestamps.
But it's simple and predictable. Since these are corner cases that we
don't expect to happen often in practice, having a less than perfect
but simple solution is preferred.